Move search use-cases to search-data
