+++ appBackend/search-data/build.gradle.kts
@@ -10,5 +10,10 @@ kotlin {
                 implementation(projects.htmlparse)
             }
         }
+        val jsTest by getting {
+            dependencies {
+                implementation(projects.searchTestfixtures)
+            }
+        }
     }
 }
