+++ appBackend/search-data/src/jsMain/kotlin/com/gchristov/thecodinglove/searchdata/SearchDataModule.kt
@@ -3,9 +3,12 @@ package com.gchristov.thecodinglove.searchdata
 import com.gchristov.thecodinglove.htmlparse.HtmlPostParser
 import com.gchristov.thecodinglove.kmpcommondi.DiModule
 import com.gchristov.thecodinglove.searchdata.model.SearchConfig
+import com.gchristov.thecodinglove.searchdata.usecase.*
 import dev.gitlive.firebase.firestore.FirebaseFirestore
 import io.ktor.client.*
+import kotlinx.coroutines.Dispatchers
 import org.kodein.di.DI
+import org.kodein.di.bindProvider
 import org.kodein.di.bindSingleton
 import org.kodein.di.instance
 
@@ -23,6 +26,24 @@ object SearchDataModule : DiModule() {
                 )
             }
             bindSingleton { provideSearchConfig() }
+            bindProvider {
+                provideSearchWithHistoryUseCase(
+                    searchRepository = instance(),
+                    searchConfig = instance()
+                )
+            }
+            bindProvider {
+                provideSearchWithSessionUseCase(
+                    searchRepository = instance(),
+                    searchWithHistoryUseCase = instance(),
+                )
+            }
+            bindProvider {
+                providePreloadSearchResultUseCase(
+                    searchRepository = instance(),
+                    searchWithHistoryUseCase = instance(),
+                )
+            }
         }
     }
 
@@ -39,4 +60,31 @@ object SearchDataModule : DiModule() {
     )
 
     private fun provideSearchConfig(): SearchConfig = SearchConfig(postsPerPage = 4)
+
+    private fun provideSearchWithHistoryUseCase(
+        searchRepository: SearchRepository,
+        searchConfig: SearchConfig
+    ): SearchWithHistoryUseCase = RealSearchWithHistoryUseCase(
+        dispatcher = Dispatchers.Default,
+        searchRepository = searchRepository,
+        searchConfig = searchConfig
+    )
+
+    private fun provideSearchWithSessionUseCase(
+        searchRepository: SearchRepository,
+        searchWithHistoryUseCase: SearchWithHistoryUseCase
+    ): SearchWithSessionUseCase = RealSearchWithSessionUseCase(
+        dispatcher = Dispatchers.Default,
+        searchRepository = searchRepository,
+        searchWithHistoryUseCase = searchWithHistoryUseCase
+    )
+
+    private fun providePreloadSearchResultUseCase(
+        searchRepository: SearchRepository,
+        searchWithHistoryUseCase: SearchWithHistoryUseCase
+    ): PreloadSearchResultUseCase = RealPreloadSearchResultUseCase(
+        dispatcher = Dispatchers.Default,
+        searchRepository = searchRepository,
+        searchWithHistoryUseCase = searchWithHistoryUseCase
+    )
 }
