+++ appBackend/search-data/src/jsMain/kotlin/com/gchristov/thecodinglove/searchdata/usecase/Random.kt
@@ -1,4 +1,4 @@
-package com.gchristov.thecodinglove.search
+package com.gchristov.thecodinglove.searchdata.usecase
 
 import arrow.core.Either
 import com.gchristov.thecodinglove.searchdata.model.Post
