+++ appBackend/search-data/src/jsMain/kotlin/com/gchristov/thecodinglove/searchdata/usecase/RealSearchWithSessionUseCase.kt
@@ -1,14 +1,11 @@
-package com.gchristov.thecodinglove.search.usecase
+package com.gchristov.thecodinglove.searchdata.usecase
 
 import arrow.core.Either
 import com.benasher44.uuid.uuid4
-import com.gchristov.thecodinglove.search.insert
 import com.gchristov.thecodinglove.searchdata.SearchException
 import com.gchristov.thecodinglove.searchdata.SearchRepository
 import com.gchristov.thecodinglove.searchdata.model.Post
 import com.gchristov.thecodinglove.searchdata.model.SearchSession
-import com.gchristov.thecodinglove.searchdata.usecase.SearchWithHistoryUseCase
-import com.gchristov.thecodinglove.searchdata.usecase.SearchWithSessionUseCase
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
