+++ appBackend/search-data/src/jsTest/kotlin/com/gchristov/thecodinglove/searchdata/usecase/RealSearchWithSessionUseCaseTest.kt
@@ -1,11 +1,9 @@
-package com.gchristov.thecodinglove.search.usecase
+package com.gchristov.thecodinglove.searchdata.usecase
 
 import arrow.core.Either
 import com.gchristov.thecodinglove.kmpcommontest.FakeCoroutineDispatcher
 import com.gchristov.thecodinglove.searchdata.SearchException
 import com.gchristov.thecodinglove.searchdata.model.SearchSession
-import com.gchristov.thecodinglove.searchdata.usecase.SearchWithHistoryUseCase
-import com.gchristov.thecodinglove.searchdata.usecase.SearchWithSessionUseCase
 import com.gchristov.thecodinglove.searchtestfixtures.*
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestResult
