+++ appBackend/search/src/jsMain/kotlin/com/gchristov/thecodinglove/search/SearchModule.kt
@@ -1,17 +1,9 @@
 package com.gchristov.thecodinglove.search
 
 import com.gchristov.thecodinglove.kmpcommondi.DiModule
-import com.gchristov.thecodinglove.search.usecase.RealPreloadSearchResultUseCase
-import com.gchristov.thecodinglove.search.usecase.RealSearchWithHistoryUseCase
-import com.gchristov.thecodinglove.search.usecase.RealSearchWithSessionUseCase
-import com.gchristov.thecodinglove.searchdata.SearchRepository
-import com.gchristov.thecodinglove.searchdata.model.SearchConfig
 import com.gchristov.thecodinglove.searchdata.usecase.PreloadSearchResultUseCase
-import com.gchristov.thecodinglove.searchdata.usecase.SearchWithHistoryUseCase
 import com.gchristov.thecodinglove.searchdata.usecase.SearchWithSessionUseCase
-import kotlinx.coroutines.Dispatchers
 import org.kodein.di.DI
-import org.kodein.di.bindProvider
 import org.kodein.di.bindSingleton
 import org.kodein.di.instance
 
@@ -20,24 +12,6 @@ object SearchModule : DiModule() {
 
     override fun bindDependencies(builder: DI.Builder) {
         builder.apply {
-            bindProvider {
-                provideSearchWithHistoryUseCase(
-                    searchRepository = instance(),
-                    searchConfig = instance()
-                )
-            }
-            bindProvider {
-                provideSearchWithSessionUseCase(
-                    searchRepository = instance(),
-                    searchWithHistoryUseCase = instance(),
-                )
-            }
-            bindProvider {
-                providePreloadSearchResultUseCase(
-                    searchRepository = instance(),
-                    searchWithHistoryUseCase = instance(),
-                )
-            }
             bindSingleton {
                 provideSearchService(
                     searchWithSessionUseCase = instance(),
@@ -46,33 +20,6 @@ object SearchModule : DiModule() {
             }
         }
     }
-
-    private fun provideSearchWithHistoryUseCase(
-        searchRepository: SearchRepository,
-        searchConfig: SearchConfig
-    ): SearchWithHistoryUseCase = RealSearchWithHistoryUseCase(
-        dispatcher = Dispatchers.Default,
-        searchRepository = searchRepository,
-        searchConfig = searchConfig
-    )
-
-    private fun provideSearchWithSessionUseCase(
-        searchRepository: SearchRepository,
-        searchWithHistoryUseCase: SearchWithHistoryUseCase
-    ): SearchWithSessionUseCase = RealSearchWithSessionUseCase(
-        dispatcher = Dispatchers.Default,
-        searchRepository = searchRepository,
-        searchWithHistoryUseCase = searchWithHistoryUseCase
-    )
-
-    private fun providePreloadSearchResultUseCase(
-        searchRepository: SearchRepository,
-        searchWithHistoryUseCase: SearchWithHistoryUseCase
-    ): PreloadSearchResultUseCase = RealPreloadSearchResultUseCase(
-        dispatcher = Dispatchers.Default,
-        searchRepository = searchRepository,
-        searchWithHistoryUseCase = searchWithHistoryUseCase
-    )
 
     private fun provideSearchService(
         searchWithSessionUseCase: SearchWithSessionUseCase,
